+++ test/test_config.py
@@ -53,6 +53,7 @@ def test_get_reports():
     expected = sorted([
         'console',
         'csv',
+        'custom',
         'json',
         'null',
         'pycodestyle',
@@ -96,7 +97,7 @@ def test_get_default_config():
     assert isinstance(actual['workers'], int)
     assert actual['workers'] >= 1
     assert actual['workers'] <= 4
-    assert actual['reports'] == [{'type': 'console'}]
+    assert actual['requested_reports'] == [{'type': 'console'}]
     assert actual['disabled'] == []
     assert actual['noqa'] == True
     assert actual['extends'] == []